:building_construction: changed creditcard payment method to new architecture
